Split observe into smaller files
